ENH: Fix Docker and Doxygen workflow actions `Node.js` warnings
Fix Docker image build and Doxygen workflow actions warnings linked to
`Node.js`: bump `actions/checkout` and `actions/upload-artifact` to v4.

Fixes:
```
build
The following actions use a deprecated Node.js version and will be forced to run on node20:
 actions/checkout@v3.
 For more info:
 https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```

```
doxygen
The following actions use a deprecated Node.js version and will be forced to run on node20:
 actions/checkout@v3, actions/upload-artifact@v3.
 For more info:
 https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```

and
```
Deprecation notice: v1, v2, and v3 of the artifact actions
The following artifacts were uploaded using a version of actions/upload-artifact that is scheduled for deprecation: "doxygen".
Please update your workflow to use v4 of the artifact actions.
Learn more: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/
```

raised for example in:
https://github.com/InsightSoftwareConsortium/ITKDoxygen/actions/runs/10436038564
and
https://github.com/InsightSoftwareConsortium/ITKDoxygen/actions/runs/10542471843